requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
joblib==1.5.2
numpy==2.3.2
scipy==1.16.1
scikit-learn==1.7.1
xgboost==3.0.5
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2

